pyproject.toml
[project]
name = "book-ai-ocr-tool"
version = "0.1.0"
description = "A tool for OCR processing of books using AI."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.60.1",
    "pillow>=10.4.0",
    "pydantic>=2.10.6",
    "google-genai>=1.56.0",
]

[project.scripts]
book-ai-ocr-tool = "book_ai_ocr_tool.book_ocr:main"
book-merge = "book_ai_ocr_tool.book_merge:main"
book-tts = "book_ai_ocr_tool.book_tts:main"
book-concat = "book_ai_ocr_tool.book_concat:main"
